Fixing closeable osmosis reader issue
